redis_prefix now only impacts one recommender
When using multiple recommenders in a script, eg for cross validation,
the redis_prefix was the same for input matrices of all recommenders.
What's more, all recommenders had access to all input matrices of all
other recommenders.
This patch fixes these problems and enables testing 2 different
recommenders in a script to see if their difference of performance is
statistically significant.